requirements.txt
firecrawl==2.7.1
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.18.1
webdriver-manager>=4.0.1